Restrict usage of the MemoryTracker constructor
A common first issue when adding new memory tracking was to create new
instances of MemoryTracker which should happen very rarely in test code.
If we create new instances outside of a MemoryTrackerManager, these
instances will not be included in the memory reporting.

The correct solution should almost always be to use an existing memory
tracker. Some common trackers to use are on both the Array and Query
classes. In cases where neither of those is appropriate, it may be time
to start considering a new instance of a memory tracker by calling
ContextResources::create_memory_tracker(). Cases where we may need to
create new instances include the usage of instrumented classes outside
of an Array or Query. I.e., when creating a new Array, there's no Array
instance which means we'll have to create a new ARRAY_CREATE (or
some other name) MemoryTracker::Type.